Fix handling of `update-alternatives`
Also make warning a little clearer using blank lines

+++ Tools/install.sh
@@ -120,7 +120,7 @@ else
             	python3-setuptools \
                 wget
 
-            $PKG_INSTALL clang-format-8 || echo "Failed installing optional clang-format-8."
+            $PKG_INSTALL clang-format-8 || printf "\nWARNING: Failed to install optional clang-format-8.\n\n"
             ;;
 
         fedora)
@@ -149,7 +149,9 @@ fi
 
 set -e
 
-sudo update-alternatives --install /usr/bin/clang-format clang-format /usr/bin/clang-format-8 100
+if [ -f "/usr/bin/clang-format-8" ]; then
+    sudo update-alternatives --install /usr/bin/clang-format clang-format /usr/bin/clang-format-8 100
+fi
 
 python3 -m pip install --upgrade pip -r $SMING_HOME/../Tools/requirements.txt
